+++ Repository/MetaRepository.php
@@ -1,0 +1,124 @@
+<?php
+
+/**
+ * Copyright (C) 2017 Rounon Dax
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace WordPress\EsiClient\Repository;
+
+use WordPress\EsiClient\Swagger;
+
+\defined('ABSPATH') or die();
+
+class MetaRepository extends Swagger {
+    /**
+     * Used ESI enpoints in this class
+     *
+     * @var array ESI enpoints
+     */
+    protected $esiEndpoints = [
+        'status_json_latest' => 'status.json?version=latest',
+        'status_json_dev' => 'status.json?version=dev',
+        'status_json_lagecy' => 'status.json?version=lagecy',
+        'status_json_meta' => 'status.json?version=meta',
+        'esi_versions' => 'versions/',
+        'esi_ping' => 'ping',
+    ];
+
+    /**
+     * Provides a general health indicator per route and method
+     *
+     * @return array of \WordPress\EsiClient\Model\Meta\Status
+     */
+    public function statusJsonLatest() {
+        $returnValue = null;
+
+        $this->setEsiMethod('get');
+        $this->setEsiRoute($this->esiEndpoints['status_json_latest']);
+        $this->setEsiVersion(null);
+
+        $esiData = $this->callEsi();
+
+        if(!\is_null($esiData)) {
+            $returnValue = $this->mapArray($esiData, '\\WordPress\EsiClient\Model\Meta\Status');
+        }
+
+        return $returnValue;
+    }
+
+    /**
+     * Provides a general health indicator per route and method
+     *
+     * @return array of \WordPress\EsiClient\Model\Meta\Status
+     */
+    public function statusJsonDev() {
+        $returnValue = null;
+
+        $this->setEsiMethod('get');
+        $this->setEsiRoute($this->esiEndpoints['status_json_dev']);
+        $this->setEsiVersion(null);
+
+        $esiData = $this->callEsi();
+
+        if(!\is_null($esiData)) {
+            $returnValue = $this->mapArray($esiData, '\\WordPress\EsiClient\Model\Meta\Status');
+        }
+
+        return $returnValue;
+    }
+
+    /**
+     * Provides a general health indicator per route and method
+     *
+     * @return array of \WordPress\EsiClient\Model\Meta\Status
+     */
+    public function statusJsonLagecy() {
+        $returnValue = null;
+
+        $this->setEsiMethod('get');
+        $this->setEsiRoute($this->esiEndpoints['status_json_lagecy']);
+        $this->setEsiVersion(null);
+
+        $esiData = $this->callEsi();
+
+        if(!\is_null($esiData)) {
+            $returnValue = $this->mapArray($esiData, '\\WordPress\EsiClient\Model\Meta\Status');
+        }
+
+        return $returnValue;
+    }
+
+    /**
+     * Provides a general health indicator per route and method
+     *
+     * @return array of \WordPress\EsiClient\Model\Meta\Status
+     */
+    public function statusJsonMeta() {
+        $returnValue = null;
+
+        $this->setEsiMethod('get');
+        $this->setEsiRoute($this->esiEndpoints['status_json_meta']);
+        $this->setEsiVersion(null);
+
+        $esiData = $this->callEsi();
+
+        if(!\is_null($esiData)) {
+            $returnValue = $this->mapArray($esiData, '\\WordPress\EsiClient\Model\Meta\Status');
+        }
+
+        return $returnValue;
+    }
+}